little changes

+++ include/task1.h
@@ -2,8 +2,8 @@
 #define TASK1_H
 #include "Circle.h"
 
-double calcDelta(double piece = 1.0, double EarthRadius = 6378100.0);
-double calcCost(double poolRadius = 3.0, double concretePrice = 1000.0,
-				double fencePrice = 2000.0, double roadWidth = 1.0);
+double calcDelta(int piece = 1, double EarthRadius = 6378100.0);
+double calcCost(double poolRadius = 3.0, int concretePrice = 1000,
+		int fencePrice = 2000, double roadWidth = 1.0);
 
 #endif
